+++ netlify/edge-functions/get-geo.tsx
@@ -1,0 +1,11 @@
+import type { Config, Context } from "@netlify/edge-functions";
+
+export default async (request: Request, context: Context) => {
+  return Response.json({
+    geo: context.geo,
+  });
+};
+
+export const config: Config = {
+  path: "/geolocation",
+};
