+++ netlify/functions/purge-cache-tag.mts
@@ -1,0 +1,22 @@
+import type { Config } from "@netlify/functions";
+import { purgeCache } from "@netlify/functions";
+
+export default async (req: Request) => {
+    const url = new URL(req.url);
+    const cacheTag = url.searchParams.get("tag");
+    if (!cacheTag) {
+        return;
+    }
+
+    console.log("Purging tag: ", cacheTag);
+
+    await purgeCache({
+        tags: [cacheTag],
+    });
+
+    return new Response("Purged!", { status: 202 });
+};
+
+export const config: Config = {
+    path: "/api/purge-cache",
+};
