+++ netlify/functions/stream.mts
@@ -1,0 +1,36 @@
+import type { Config } from "@netlify/functions";
+
+export default async () => {
+    const encoder = new TextEncoder();
+    const formatter = new Intl.DateTimeFormat("en", { timeStyle: "medium" });
+
+    const body = new ReadableStream({
+        start(controller) {
+            controller.enqueue(encoder.encode("<html><body><ol>"));
+            let i = 0;
+            const timer = setInterval(() => {
+                controller.enqueue(
+                    encoder.encode(
+                        `<li>Hello at ${formatter.format(new Date())}</li>\n\n`,
+                    ),
+                );
+                if (i++ >= 5) {
+                    controller.enqueue(encoder.encode("</ol></body></html>"));
+                    controller.close();
+                    clearInterval(timer);
+                }
+            }, 1000);
+        },
+    });
+
+    return new Response(body, {
+        headers: {
+            "content-type": "text/html",
+        },
+        status: 200,
+    });
+};
+
+export const config: Config = {
+    path: "/api/stream",
+};
